Correctly track intermediate files produced during collectstatic
Unfortunately not all intermediate files are yielded by the
`post_process` method, so the only way to do this is to overide the
`hashed_name` method and store them in an instance attribute.

This also means that we defer all compression to the end of the
post-process run, as we don't know until then which are the final files
that need compressing.

Closes #147